14/06 21:02

+++ src/components/AddressItem/index.js
@@ -4,18 +4,19 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faXmark } from '@fortawesome/free-solid-svg-icons';
 import { useNavigate } from 'react-router-dom';
 import { initStateAddress, addressReducer } from '~/reducers/addressReducers';
-import { setIDAccount, setInfoPhone, setInfoName, setAddress } from '~/actions/addressActions';
+import { setIDAccount, setInfoPhone, setInfoName, setAddress, setIDInfo } from '~/actions/addressActions';
 
 import classNames from 'classnames/bind';
 import styles from './AddressItem.module.scss';
 
 const cx = classNames.bind(styles);
 
-function AddressItem({ ID_GIAOHANG, ID_TAIKHOAN, TENNGUOINHAN, SDT, TENDIACHI }) {
+function AddressItem({ ID_GIAOHANG, ID_TAIKHOAN, TENNGUOINHAN, SDT, TENDIACHI, nonUpdate }) {
     const [statusModal, setStatusModal] = useState(false);
     const [stateAddress, dispatchAddress] = useReducer(addressReducer, initStateAddress);
     useEffect(() => {
         dispatchAddress(setIDAccount(ID_TAIKHOAN));
+        dispatchAddress(setIDInfo(ID_GIAOHANG));
         dispatchAddress(setInfoName(TENNGUOINHAN));
         dispatchAddress(setInfoPhone(SDT));
         dispatchAddress(setAddress(TENDIACHI));
@@ -27,36 +28,68 @@ function AddressItem({ ID_GIAOHANG, ID_TAIKHOAN, TENNGUOINHAN, SDT, TENDIACHI })
     const hideBuyTickets = () => {
         setStatusModal(false);
     };
-
-    const deleteAddress = async () => {
+    const handleSubmitUpdate = async (e) => {
+        e.preventDefault();
+        await updateAddress({
+            stateAddress,
+        });
+    };
+    const updateAddress = async () => {
         await axios
-            .post('http://26.17.209.162/api/shippinginfo/post', {
-                type: 'delete',
-                data: { ID_TAIKHOAN: ID_TAIKHOAN, ID_GIAOHANG: ID_GIAOHANG },
+            .post('http://26.87.217.216:8080/api/giaohang/post', {
+                type: 'update',
+                data: stateAddress,
             })
             .then((res) => {
-                if ((res.data != 0) & (res.data != -1)) {
-                    alert('Xóa địa chỉ thành công');
+                if (res.data == 1) {
+                    alert('Cập nhật địa chỉ thành công');
                     window.location.reload();
+                } else if (res.data == -1) {
+                    alert('Cập nhật địa chỉ thất bại');
                 }
             });
     };
-    return (
-        <div className={cx('wrapper')}>
-            <div className={cx('inner')}>
-                <div className={cx('name')}>{TENNGUOINHAN}</div>
-                <div className={cx('address')}>{TENDIACHI}</div>
-                <div className={cx('phone')}>{SDT}</div>
 
-                <div className={cx('action')}>
-                    <button className={cx('update_btn')} onClick={showBuyTickets}>
-                        Cập nhật
-                    </button>
-                    <button className={cx('delete_btn')} onClick={deleteAddress}>
-                        Xóa
-                    </button>
+    const deleteAddress = async () => {
+        if (window.confirm('Bạn có chắc chắn muốn xóa địa chỉ này?')) {
+            await axios
+                .post('http://26.87.217.216:8080/api/giaohang/post', {
+                    type: 'delete',
+                    data: { ID_TAIKHOAN: ID_TAIKHOAN, ID_GIAOHANG: ID_GIAOHANG },
+                })
+                .then((res) => {
+                    if (res.data == 1) {
+                        alert('Xóa địa chỉ thành công');
+                        window.location.reload();
+                    } else if (res.data == -1) {
+                        alert('Xóa địa chỉ thất bại');
+                    }
+                });
+        }
+    };
+    return (
+        <>
+            <div className={cx('wrapper')}>
+                <div className={cx('inner')}>
+                    <div className={cx('name')}>{TENNGUOINHAN}</div>
+                    <div className={cx('address')}>{TENDIACHI}</div>
+                    <div className={cx('phone')}>{SDT}</div>
+
+                    <div className={cx('action')}>
+                        {nonUpdate ? (
+                            ''
+                        ) : (
+                            <button className={cx('update_btn')} onClick={showBuyTickets}>
+                                Cập nhật
+                            </button>
+                        )}
+                        <button className={cx('delete_btn')} onClick={deleteAddress}>
+                            Xóa
+                        </button>
+                    </div>
                 </div>
             </div>
+
             {/* Begin modal */}
             <div className={cx('modal', statusModal ? 'open' : '')} onClick={hideBuyTickets}>
                 <div
@@ -70,7 +103,7 @@ function AddressItem({ ID_GIAOHANG, ID_TAIKHOAN, TENNGUOINHAN, SDT, TENDIACHI })
                         <FontAwesomeIcon className={cx('modal--close')} icon={faXmark} onClick={hideBuyTickets} />
                     </div>
 
-                    <form>
+                    <form onSubmit={handleSubmitUpdate}>
                         <div className={cx('stock-list')}>
                             <div className={cx('info')}>
                                 <label htmlFor="" className={cx('input-label')}>
@@ -101,11 +134,18 @@ function AddressItem({ ID_GIAOHANG, ID_TAIKHOAN, TENNGUOINHAN, SDT, TENDIACHI })
                                     Số điện thoại
                                 </label>
                                 <input
+                                    maxLength={10}
                                     className={cx('input-item')}
                                     type="text"
                                     value={stateAddress.SDT ? stateAddress.SDT : null}
                                     placeholder="Số điện thoại"
+                                    name="phone"
                                     onChange={(e) => dispatchAddress(setInfoPhone(e.target.value))}
+                                    onKeyPress={(event) => {
+                                        if (!/[0-9]/.test(event.key)) {
+                                            event.preventDefault();
+                                        }
+                                    }}
                                 />
                             </div>
                         </div>
@@ -114,7 +154,7 @@ function AddressItem({ ID_GIAOHANG, ID_TAIKHOAN, TENNGUOINHAN, SDT, TENDIACHI })
                 </div>
             </div>
             {/* End modal */}
-        </div>
+        </>
     );
 }
